refactor(notification): extract utils, hooks, config from NotificationContent
- utils/notificationHelpers.ts: formatBytes, formatEta, normalizeUrl, resolveNotificationIconUrl, normalizeUpdateSource, isProOnlySource, sanitizeFavorites, persistFavorites
- config/notificationConstants.ts: 4 store key constants
- config/notificationTypes.ts: UrlFavoriteItem, UpdateSourceKey, DownloadProgressData, NotificationContentProps
- hooks/useNotificationFavorites.ts: favorites loading and setter
- hooks/useUpdateDownloadProgress.ts: update download progress listener
- NotificationContent.tsx: simplified from 759 to ~530 lines
